Change deleteEmployee API

+++ API Code/API-Functions.js
@@ -97,17 +97,23 @@ const createNewEmployee = async (comp_id, first_name, last_name, email, activati
 const deleteEmployee = async (employee_id) => {
     const db = await pool.connect();
     try {
-        const results = await db.query(`DELETE FROM employeesincompany WHERE employee_id = $1`, [employee_id]);
-		console.log('Results from employeesincompany deletion:', results);
-        if (results.rowCount > 0 ) {
-			const results2 = await db.query(`DELETE FROM employee WHERE id = $1`, [employee_id]);
-			if (results2.rowCount > 0 ) {
-            	console.log('Delete successful');
-            	return true;
-        	} else {
-				console.log('No rows were deleted2');
-				return false;
-        	}
+		const scheduleResult = await db.query(`DELETE FROM schedulefromemployee WHERE employee_id = $1`,[employee_id]);
+		console.log("BLBLBLLBLBLBL");
+		if (scheduleResult.rowCount > 0)
+		{
+			const results = await db.query(`DELETE FROM employeesincompany WHERE employee_id = $1`, [employee_id]);
+			console.log('Results from employeesincompany deletion:', results);
+			if (results.rowCount > 0 )
+			{
+				const results2 = await db.query(`DELETE FROM employee WHERE id = $1`, [employee_id]);
+				if (results2.rowCount > 0 ) {
+					console.log('Delete successful');
+					return true;
+				} else {
+					console.log('No rows were deleted2');
+					return false;
+				}
+			}
 	 	}
 		else {
 			console.log('No rows were deleted');
@@ -120,6 +126,8 @@ const deleteEmployee = async (employee_id) => {
         db.release(); 
     }
 };
+
+// deleteEmployee(1);
 
 // Functions to fetch data from the database
 const getAllEmployeeData = async () => {
